+++ packages/orq-rc/RELEASES.md
@@ -8839,3 +8839,13 @@ Based on:
 - [typescript v4.9.0-rc.21] packages/orq-rc
 ### Releases
 - [NPM v4.9.0-rc.21] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.21 - packages/orq-rc
+
+## 2026-05-04 15:34:18
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.11 (2.881.17) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.9.0-rc.22] packages/orq-rc
+### Releases
+- [NPM v4.9.0-rc.22] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.22 - packages/orq-rc